binance-stream-receiver/requirements.txt
gevent==21.1.2
greenlet==1.1.0
pymongo==3.11.4
websocket-client==1.1.0
zope.event==4.5.0
zope.interface==5.4.0
